bytes: move is_zero utility into storage parser
The bytes/utils.h::is_zero is moved into its only used location,
storage::parser. A useful generalization would be an iobuf version of
memcmp, but is_zero is specific and noisy in the bytes library.
