+++ src/v/storage/parser_utils.h
@@ -60,4 +60,25 @@ compress_batch(model::compression, const model::record_batch&);
 /// \brief resets the size, header crc and payload crc
 void reset_size_checksum_metadata(model::record_batch_header&, const iobuf&);
 
+inline bool is_zero(const char* data, size_t size) {
+    // NOLINTNEXTLINE(cppcoreguidelines-pro-bounds-pointer-arithmetic)
+    return data[0] == 0 && memcmp(data, data + 1, size - 1) == 0;
+}
+
+inline bool is_zero(const iobuf& buffer) {
+    if (buffer.empty()) {
+        return false;
+    }
+    bool ret = true;
+    iobuf::iterator_consumer in(buffer.cbegin(), buffer.cend());
+    in.consume(buffer.size_bytes(), [&ret](const char* src, size_t len) {
+        if (!is_zero(src, len)) {
+            ret = false;
+            return ss::stop_iteration::yes;
+        }
+        return ss::stop_iteration::no;
+    });
+    return ret;
+}
+
 } // namespace storage::internal
